func/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage==1.14.0